Move stop hook runtime into hook helpers
